Use certbot to update certificates
For each domainname passed, attempt to create a single certificate with
both the domainname as well as `*.domainname` (e.g., 'example.com' and
'*.example.com').  Don't bother making the user add pipe-delimited
markup for wildcards on a single domain.  If we need to do so, we can add
a flag to inhibit this behavior later.

Next, if no domainnames are passed, just renew the existing certificates.

Then, we added DEBUGFLAGS and set it to a default value that just does
a dry run for test certificates.  This will likely be reversed later or
we can add flags.

What's still left to do is to build fullchain certificates by ourselves.